Stop keepalive on delete requests.
This isn't a particularly nice solution, but currently the delete handler is really very dumb.
It parses the HTTP response _just_ to get the error code, and leaves any extra headers on the
socket. This means that with keepalive connections, subsequent requests will end up trying to
read in the headers for previous requests, which makes everything go horribly wrong.

The proper fix to this would be to use a proper HTTP client here, rather than a horrible hack.